This commit reverts 635ac57
To be looked at later.
As per EyalAr:
“I like the code to be consistent, so if we leave it like this, we have
to make similar changes in other places in the code.
I agree that this code gives us more flexibility regarding the number
of channels, but we need to design those changes properly, as they
affect other parts of the library.
Let's revert for now, and later decide how we properly bring this back
across the library.”